Modified type for argument to lms_socket_create() in accordance with new policy.
git-svn-id: https://svn.code.sf.net/p/kageki/code/libmsocket@20 10193a71-e229-423b-a074-d0e0c923123e

+++ msocket-internal.h
@@ -346,7 +346,7 @@ extern int lms_proxyclear(void);
 extern int *lms_socket_init(void);
 extern void lms_socket_insertfd(MSocket *m);
 extern inline MSocket *lms_socket_findbyfd(int fd);
-extern MSocket *lms_socket_create(int type);
+extern MSocket *lms_socket_create(int16_t type);
 extern int lms_socket_close(MSocket *ptr);
 extern int lms_socket_destroy(MSocket *ptr);
 extern unsigned int lms_socket_destroytype(unsigned short type, short killad);
