+++ msocket.h
@@ -240,7 +240,7 @@ extern int lms_proxyclear(void);
 /* socket.c */
 extern void lms_socket_insertfd(MSocket *m);
 extern inline MSocket *lms_socket_findbyfd(int fd);
-extern MSocket *lms_socket_create(int type);
+extern MSocket *lms_socket_create(int16_t type);
 extern int lms_socket_close(MSocket *ptr);
 extern int lms_socket_destroy(MSocket *ptr);
 extern unsigned int lms_socket_destroytype(unsigned short type, short killad);
